Feat: Rearrange repo as a monorepo
