+++ .eslintrc.json
@@ -70,6 +70,7 @@
     "space-in-parens": ["off"],
     "space-infix-ops": ["off"],
     "spaced-comment": ["off"],
-    "vars-on-top": ["off"]
+    "vars-on-top": ["off"],
+    "import/extensions": ["off"]
   }
 }
