+++ packages/client/src/index.js
@@ -1,0 +1,6 @@
+'use strict';
+
+import OpenReviewClient from './client.js';
+import Tools from './tools.js';
+
+export { OpenReviewClient, Tools };
